json: automatically scan title when processing json jobs
Simplifies the WinGui.

This also changes how jobs are processed.  Creating the sub-jobs for
multiple passes is delayed until after scanning and immediately before
running the job.

Working status has also changed.  Sub-job passes are identified in status
with an ID that allows the frontend to definitively identify what pass
is in progress.


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6976 b64f7644-9d1e-0410-96f1-a4d463321fa5